refactor(ndk): move event writing funcs into own file
Similar to the creation of serializer/event_reader.c, this moves all
async-signal-safe functions for writing an event into their own file.
The serializer.c file is now a frontend to event reading and writing,
while organizing the helper functions as their own units. A future
changeset will do something comparable with the json string writer
functions.
