+++ bugsnag-plugin-android-ndk/src/main/CMakeLists.txt
@@ -15,7 +15,9 @@ add_library( # Specifies the name of the library.
     jni/handlers/signal_handler.c
     jni/handlers/cpp_handler.cpp
     jni/utils/crash_info.c
+    jni/utils/serializer/buffered_writer.c
     jni/utils/serializer/event_reader.c
+    jni/utils/serializer/event_writer.c
     jni/utils/stack_unwinder.c
     jni/utils/stack_unwinder_libunwindstack.cpp
     jni/utils/stack_unwinder_libcorkscrew.c
@@ -24,7 +26,6 @@ add_library( # Specifies the name of the library.
     jni/utils/serializer.c
     jni/utils/string.c
     jni/utils/threads.c
-    jni/utils/buffered_writer.c
     jni/deps/parson/parson.c
              )
