+++ bugsnag-plugin-android-ndk/src/main/jni/utils/serializer/buffered_writer.h
